adjusting about section in tutorials
